src/data_population/requirements.txt
mysql-connector-python==8.0.32
redis==4.3.3
fastapi
uvicorn